+++ server/run/test_run.py
@@ -149,19 +149,19 @@ def parse_all_cases(target_exec_dir, target, kind_bench, bench_name,
                                         parser, subsection_file,
                                         tmp_parser_file)
             else:
-		if re.search('application', kind_bench) and bench_name == "nginx":
-                    outfp = open(logfile, 'r')
+		if bench_name == "nginx":
                     infp = open(tmp_log_file, 'a+')
+                    outfp = open(logfile, 'r')
                     infp.write(re.findall("test start\s+%+(.*?)%+\s+test_end", outfp.read(), re.DOTALL)[i])
 		    no_of_clients = configRun.get(sections_run[i], 'no_of_clients')
 		    for j in range(1, int(no_of_clients) + 1):
 			weighttp_log_file = Folder.exec_dir + "/" + "weighttp_client_" + str(j) + "_output.log"
-                    	outfp_weighttp = open(weighttp_log_file, 'r')
-                        content = re.findall("test start\s+%+(.*?)%+\s+test_end", outfp_weighttp.read(), re.DOTALL)[i]
-			infp.write(content)
-                    infp.close()
-                    infp = open(tmp_log_file, 'r')
-		    content = infp.read()
+			file_present = os.path.isfile(weighttp_log_file)
+			if file_present == True:
+                    	    outfp_weighttp = open(weighttp_log_file, 'r')
+                            content = re.findall("test start\s+%+(.*?)%+\s+test_end", outfp_weighttp.read(), re.DOTALL)[i]
+			    infp.write(content)
+                    
                     infp.close()
                     outfp.close()
                     parser_result = parser_case(kind_bench, bench_name, parser_file,
@@ -285,19 +285,18 @@ def run_all_cases(target_exec_dir, target, kind_bench, bench_name,
 
     nginx_tmp_log_file = None
 
-    if re.search('application', kind_bench):
-        if bench_name == "nginx":
-	    nginx_log_file = {}
-	    nginx_tmp_log_file = {}
-            no_of_clients = len(nginx_clients)
-            for i in range(1, no_of_clients+1):
-                filename = "weighttp_client_" + str(i)
-                nginx_log_bench = os.path.join(Folder.exec_dir, filename)
-                nginx_log_file[str(i)] = nginx_log_bench + "_output.log"
-                nginx_tmp_log_file[str(i)] = nginx_log_bench + "_output_tmp.log"
+    if re.search('application', kind_bench) and bench_name == "nginx":
+	nginx_log_file = {}
+	nginx_tmp_log_file = {}
+        no_of_clients = len(nginx_clients)
+        for i in range(1, no_of_clients+1):
+            filename = "weighttp_client_" + str(i)
+            nginx_log_bench = os.path.join(Folder.exec_dir, filename)
+            nginx_log_file[str(i)] = nginx_log_bench + "_output.log"
+            nginx_tmp_log_file[str(i)] = nginx_log_bench + "_output_tmp.log"
 
-                if os.path.exists(nginx_log_file[str(i)]):
-                    os.remove(nginx_log_file[str(i)])
+            if os.path.exists(nginx_log_file[str(i)]):
+                os.remove(nginx_log_file[str(i)])
 
     starttime = datetime.datetime.now()
     if os.path.exists(Folder.caliper_log_file):
@@ -360,29 +359,27 @@ def run_all_cases(target_exec_dir, target, kind_bench, bench_name,
         if os.path.exists(tmp_log_file):
             os.remove(tmp_log_file)
 
-	if re.search('application', kind_bench):
-            if bench_name == "nginx":
-                no_of_clients = len(nginx_clients)
-            	for j in range(1, no_of_clients+1):
-            	    if os.path.exists(nginx_tmp_log_file[str(j)]):
-                        os.remove(nginx_tmp_log_file[str(j)])
+	if re.search('application', kind_bench) and bench_name == "nginx":
+            no_of_clients = len(nginx_clients)
+            for j in range(1, no_of_clients+1):
+                if os.path.exists(nginx_tmp_log_file[str(j)]):
+                    os.remove(nginx_tmp_log_file[str(j)])
 
         server_run_command = get_server_command(kind_bench, sections_run[i])
 
         nginx_clients_count = None
 	client_command_dic = None
-        if re.search('application', kind_bench):
-	    if bench_name == "nginx":
-		no_of_clients = configRun.get(sections_run[i], 'no_of_clients')
-		if len(nginx_clients) >= int(no_of_clients):
-		    client_command_dic = {}
-		    for j in range(1, int(no_of_clients)+1):
-			client_command = "command" + str(j)
-			client_command_dic[str(j)] = get_nginx_client_command(kind_bench, sections_run[i], client_command)
-		    nginx_clients_count = int(no_of_clients)
-		else:
-		    logging.info("Please specify client in the client config file")
-		    continue
+        if re.search('application', kind_bench) and bench_name == "nginx":
+	    no_of_clients = configRun.get(sections_run[i], 'no_of_clients')
+	    if len(nginx_clients) >= int(no_of_clients):
+	        client_command_dic = {}
+	        for j in range(1, int(no_of_clients)+1):
+		    client_command = "command" + str(j)
+		    client_command_dic[str(j)] = get_nginx_client_command(kind_bench, sections_run[i], client_command)
+	        nginx_clients_count = int(no_of_clients)
+	    else:
+	        logging.info("Please specify client in the client config file")
+		continue
       
         logging.debug("Get the server command is: %s" % server_run_command)
         # run the command of the benchmarks
@@ -401,13 +398,12 @@ def run_all_cases(target_exec_dir, target, kind_bench, bench_name,
             if os.path.exists(tmp_log_file):
                 os.remove(tmp_log_file)
 
-	    if re.search('application', kind_bench):
-        	if bench_name == "nginx":
-            	    no_of_clients = len(nginx_clients)
-            	    for j in range(1, no_of_clients + 1):
-            		server_utils.file_copy(nginx_log_file[str(j)], nginx_tmp_log_file[str(j)], 'a+')
-            		if os.path.exists(nginx_tmp_log_file[str(j)]):
-                	    os.remove(nginx_tmp_log_file[str(j)])
+	    if re.search('application', kind_bench) and bench_name == "nginx":
+	    	no_of_clients = configRun.get(sections_run[i], 'no_of_clients')
+            	for j in range(1, int(no_of_clients) + 1):
+            	    server_utils.file_copy(nginx_log_file[str(j)], nginx_tmp_log_file[str(j)], 'a+')
+            	    if os.path.exists(nginx_tmp_log_file[str(j)]):
+                        os.remove(nginx_tmp_log_file[str(j)])
 
             run_flag = server_utils.get_fault_tolerance_config(
                                 'fault_tolerance', 'run_error_continue')
@@ -423,22 +419,20 @@ def run_all_cases(target_exec_dir, target, kind_bench, bench_name,
                     server_utils.file_copy(tmp_log_file,subsection_file, 'a+')
             server_utils.file_copy(logfile, tmp_log_file, 'a+')
 
-	    if re.search('application', kind_bench):
-        	if bench_name == "nginx":
-            	    no_of_clients = len(nginx_clients)
-            	    for j in range(1, no_of_clients + 1):
-            		server_utils.file_copy(nginx_log_file[str(j)], nginx_tmp_log_file[str(j)], 'a+')
+	    if re.search('application', kind_bench) and bench_name == "nginx":
+	    	no_of_clients = configRun.get(sections_run[i], 'no_of_clients')
+            	for j in range(1, int(no_of_clients) + 1):
+            	    server_utils.file_copy(nginx_log_file[str(j)], nginx_tmp_log_file[str(j)], 'a+')
 
             if flag != 1:
                 logging.info("There is wrong when running the command \"%s\""
                                 % command)
 
-	    	if re.search('application', kind_bench):
-        	    if bench_name == "nginx":
-            	        no_of_clients = len(nginx_clients)
-            	        for j in range(1, no_of_clients + 1):
-            		    if os.path.exists(nginx_tmp_log_file[str(j)]):
-                	        os.remove(nginx_tmp_log_file[str(j)])
+	    	if re.search('application', kind_bench) and bench_name == "nginx":
+	    	    no_of_clients = configRun.get(sections_run[i], 'no_of_clients')
+            	    for j in range(1, int(no_of_clients) + 1):
+            	        if os.path.exists(nginx_tmp_log_file[str(j)]):
+                	    os.remove(nginx_tmp_log_file[str(j)])
 
                 if os.path.exists(tmp_log_file):
                     os.remove(tmp_log_file)
@@ -454,12 +448,11 @@ def run_all_cases(target_exec_dir, target, kind_bench, bench_name,
             if os.path.exists(tmp_log_file):
                 os.remove(tmp_log_file)
 
-	    if re.search('application', kind_bench):
-       	    	if bench_name == "nginx":
-            	    no_of_clients = len(nginx_clients)
-            	    for j in range(1, no_of_clients + 1):
-            		if os.path.exists(nginx_tmp_log_file[str(j)]):
-                	    os.remove(nginx_tmp_log_file[str(j)])
+	    if re.search('application', kind_bench) and bench_name == "nginx":
+	    	no_of_clients = configRun.get(sections_run[i], 'no_of_clients')
+            	for j in range(1, int(no_of_clients) + 1):
+            	    if os.path.exists(nginx_tmp_log_file[str(j)]):
+                        os.remove(nginx_tmp_log_file[str(j)])
 
     endtime = datetime.datetime.now()
     result = subprocess.call("echo '$$ %s EXECUTION STOP: %s' >> %s"
@@ -596,7 +589,7 @@ def run_server_command(cmd_sec_name, commands, tmp_logfile, kind_bench, server, 
         # the commands is multiple lines, and was included by Quotation
 	final_commands = get_actual_commands(commands, server)
         if final_commands is not None and final_commands != '':
-            logging.info("the actual commands running on the remote client is: %s" % final_commands)
+            logging.debug("the actual commands running on the remote client is: %s" % final_commands)
             [out, returncode] = run_remote_server_commands(final_commands, server, fp, fp, timeout)
         else:
             return ['Not command specified', -1]
@@ -633,7 +626,7 @@ def run_remote_client_commands(exec_dir, kind_bench, commands, target,
         #the commands is multiple lines, and was included by Quotation
         final_commands = remote_commands_deal(commands, exec_dir, target)
         if final_commands is not None and final_commands != '':
-            logging.info("the actual commands running on the remote target ---------"
+            logging.debug("the actual commands running on the remote target"
                             "is: %s" % final_commands)
             result = target.run(final_commands, stdout_tee=stdout_tee,
                                 stderr_tee=stderr_tee, verbose=True)
@@ -692,7 +685,7 @@ def run_client_command(cmd_sec_name, tmp_logfile, kind_bench,
             [out, returncode] = run_commands(host_exec_dir, kind_bench,
                                                 command, fp, fp)
         else:
-            logging.info("client command in remote target is: %s" % command)
+            logging.debug("client command in remote target is: %s" % command)
             [out, returncode] = run_remote_client_commands(host_exec_dir, kind_bench,
                                                     command, target, fp, fp)
     except error.ServRunError, e:
@@ -721,6 +714,41 @@ def client_thread_func(cmd_sec_name, server_run_command, tmp_logfile,
     flag = run_server_command(cmd_sec_name, server_run_command, tmp_logfile,
                     kind_bench, server, 5000)
 
+def get_count(host_login):
+    p1 = subprocess.Popen(['ssh','%s' % host_login, 'ps','-ef'], stdout=subprocess.PIPE)
+    p2 = subprocess.Popen(['grep', '-c', 'run_weighttp_script.sh'], stdin=p1.stdout, stdout=subprocess.PIPE)
+    data = p2.communicate()
+    data = str(data[0]).split("\n")
+    return data[0]
+
+def stop_weighttp_client(nginx_clients_count):
+    filename = caliper_path.folder_ope.workspace + "/config" + "/client_config.cfg"
+    fp = open(filename,"r")
+    content = fp.read()
+
+    for i in range(1, nginx_clients_count + 1):
+      	temp_ip = re.findall("(client\_%d\_ip):.*?" % i, content)
+      	temp_user = re.findall("(client\_%d\_user):.*?" % i, content)
+      	ip = temp_ip[0]
+      	user = temp_user[0]
+       	client_ip = settings.get_value('nginx', ip , type=str)
+       	client_user = settings.get_value('nginx', user , type=str)
+	host_login = client_user + "@" + client_ip
+
+        process_count = get_count(host_login)
+        if process_count > 0:
+            for j in range(0, int(process_count)):
+                p1 = subprocess.Popen(['ssh', '%s' % host_login, 'ps','-ef'], stdout=subprocess.PIPE)
+                p2 = subprocess.Popen(['grep','run_weighttp_script.sh'], stdin=p1.stdout, stdout=subprocess.PIPE)
+                string_val = "BEGIN {}/bash/"
+                p3 = subprocess.Popen(['awk', '%s' % string_val], stdin=p2.stdout, stdout=subprocess.PIPE)
+                p4 = subprocess.Popen(['awk', '{print $8} {print $2}'], stdin=p3.stdout, stdout=subprocess.PIPE)
+                data = p4.communicate()
+                bash_data = str(data[0]).split("\n")[0]
+                bash_pid = str(data[0]).split("\n")[1]
+                p1 = subprocess.Popen(['ssh', '%s' % host_login, 'kill','-9', bash_pid], stdout=subprocess.PIPE)
+    fp.close()
+
 def run_kinds_commands(cmd_sec_name, server_run_command, tmp_logfile,
                         kind_bench, bench_name, target, command, server, nginx_clients=None, 
 			client_command_dic=None, nginx_clients_count=None, nginx_tmp_log_file=None):
@@ -741,11 +769,12 @@ def run_kinds_commands(cmd_sec_name, server_run_command, tmp_logfile,
 	        threadname = myThread(i, cmd_sec_name, client_command_dic[str(i)], nginx_tmp_log_file[str(i)],
                            kind_bench, nginx_clients[str(i)])
         	threadname.start()
-		logging.info(" %d command runs............" % i)
 
 	    flag = run_client_command(cmd_sec_name, tmp_logfile, kind_bench,
                                     target, command)
-	    logging.info("target command runs............")
+
+            stop_weighttp_client(nginx_clients_count)
+
 	else:
             flag = run_client_command(cmd_sec_name, tmp_logfile, kind_bench,
                                     target, command)
@@ -801,9 +830,7 @@ def parser_case(kind_bench, bench_name, parser_file, parser, infile, outfile):
                     # call the parser function to filter the output
                     logging.debug("Begining to parser the result of the case")
                     result = methodToCall(contents, outfp)
-		    print result
                 except Exception, e:
-		    logging.info("dhfhdjkdfdkfdkfdkf %s" % e)
                     logging.info(e)
                     return -5
 	    else:
